# src/api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "qemat",
      "description": "qemat main backend endpoints testing",
      "contact": {}
    },
    "host": "https://qemat-staging.herokuapp.com/api",
    "basePath": "/",
    "securityDefinitions": {},
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/histogram/generateStats": {
        "get": {
          "description": "Login with tokenid from firebase, this will check if it already exists in the system, if not it will use this to login to firebase, if the idtoken is correct it will store it in the system for further use",
          "summary": "GenerateStats (TEMP)",
          "tags": [
            "Temp"
          ],
          "operationId": "HistogramGenerateStatsGet",
          "deprecated": false,
          "produces": [
            "application/json; charset=utf-8"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/GetAllScatterPlotSuccess"
                }
              },
              "headers": {
                "Access-Control-Allow-Origin": {
                  "type": "string",
                  "default": "*"
                },
                "Strict-Transport-Security": {
                  "type": "string",
                  "default": "max-age=15552000; includeSubDomains"
                },
                "X-DNS-Prefetch-Control": {
                  "type": "string",
                  "default": "off"
                },
                "X-Frame-Options": {
                  "type": "string",
                  "default": "SAMEORIGIN"
                },
                "X-Download-Options": {
                  "type": "string",
                  "default": "noopen"
                },
                "X-Content-Type-Options": {
                  "type": "string",
                  "default": "nosniff"
                },
                "X-XSS-Protection": {
                  "type": "string",
                  "default": "1; mode=block"
                },
                "Surrogate-Control": {
                  "type": "string",
                  "default": "no-store"
                },
                "Cache-Control": {
                  "type": "string",
                  "default": "no-store, no-cache, must-revalidate, proxy-revalidate"
                },
                "Pragma": {
                  "type": "string",
                  "default": "no-cache"
                },
                "Expires": {
                  "type": "string",
                  "default": "0"
                },
                "ETag": {
                  "type": "string",
                  "default": "W/\"8fdc-f0ODwLNa3sq+WYyHSYTCY76wlnQ\""
                },
                "Vary": {
                  "type": "string",
                  "default": "Accept-Encoding"
                },
                "Content-Encoding": {
                  "type": "string",
                  "default": "gzip"
                },
                "Date": {
                  "type": "string",
                  "default": "Mon, 08 Apr 2019 09:11:01 GMT"
                },
                "Connection": {
                  "type": "string",
                  "default": "keep-alive"
                },
                "Transfer-Encoding": {
                  "type": "string",
                  "default": "chunked"
                }
              }
            }
          }
        }
      },
...
      "/histogram/get": {
        "get": {
          "description": "Get Histogram based of make, model, year, battery.",
          "summary": "Get Histogram",
          "tags": [
            "Histogram"
          ],
          "operationId": "HistogramGetGet",
          "deprecated": false,
          "produces": [
            "application/json; charset=utf-8"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            },
            {
              "name": "year",
              "in": "query",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": "Year of the car released"
            },
            {
              "name": "make",
              "in": "query",
              "required": true,
              "type": "string",
              "description": "Maker of the car"
            },
            {
              "name": "model",
              "in": "query",
              "required": true,
              "type": "string",
              "description": "Model of the car"
            },
            {
              "name": "battery",
              "in": "query",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": "Battery of the car."
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/GetHistogramSuccess"
                }
              },
              "examples": {
                "application/json; charset=utf-8": [
                  {
                    "priceUpper": 50000,
                    "priceLower": 40000,
                    "count": 26
                  }
                ]
              },
              "headers": {
                "Access-Control-Allow-Origin": {
                  "type": "string",
                  "default": "*"
                },
                "Strict-Transport-Security": {
                  "type": "string",
                  "default": "max-age=15552000; includeSubDomains"
                },
                "X-DNS-Prefetch-Control": {
                  "type": "string",
                  "default": "off"
                },
                "X-Frame-Options": {
                  "type": "string",
                  "default": "SAMEORIGIN"
                },
                "X-Download-Options": {
                  "type": "string",
                  "default": "noopen"
                },
                "X-Content-Type-Options": {
                  "type": "string",
                  "default": "nosniff"
                },
                "X-XSS-Protection": {
                  "type": "string",
                  "default": "1; mode=block"
                },
                "Surrogate-Control": {
                  "type": "string",
                  "default": "no-store"
                },
                "Cache-Control": {
                  "type": "string",
                  "default": "no-store, no-cache, must-revalidate, proxy-revalidate"
                },
                "Pragma": {
                  "type": "string",
                  "default": "no-cache"
                },
                "Expires": {
                  "type": "string",
                  "default": "0"
                },
                "Content-Length": {
                  "type": "string",
                  "default": "52"
                },
                "ETag": {
                  "type": "string",
                  "default": "W/\"34-N8fmf40ubTjIT6DJACIOt9cx0ew\""
                },
                "Vary": {
                  "type": "string",
                  "default": "Accept-Encoding"
                },
                "Date": {
                  "type": "string",
                  "default": "Wed, 17 Apr 2019 12:50:02 GMT"
                },
                "Connection": {
                  "type": "string",
                  "default": "keep-alive"
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/RegisterVehicleUnauthorised"
              },
              "examples": {
                "application/json": {
                  "name": "AuthorizationRequiredError",
                  "message": "Authorization is required for request on GET /api/histogram/get?year=2013&make=tesla&model=Tesla%20Model%20S&battery=60",
                  "errors": []
                }
              }
            }
          }
        }
      },
      "/histogram/getAll": {
        "get": {
          "description": "Get Histogram data for all the entries. Right now this is recommended.",
          "summary": "Get All Histogram data",
          "tags": [
            "Histogram"
          ],
          "operationId": "HistogramGetAllGet",
          "deprecated": false,
          "produces": [
            "application/json; charset=utf-8"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/GetAllHistogramdataSuccess"
                }
              },
              "examples": {
                "application/json; charset=utf-8": [
                  {
                    "priceUpper": 30000,
                    "priceLower": 20000,
                    "count": 1
                  },
                  {
                    "priceUpper": 40000,
                    "priceLower": 30000,
                    "count": 5
                  },
                  {
                    "priceUpper": 70000,
                    "priceLower": 60000,
                    "count": 5
                  },
                  {
                    "priceUpper": 50000,
                    "priceLower": 40000,
                    "count": 26
                  },
                  {
                    "priceUpper": 50000,
                    "priceLower": 40000,
                    "count": 40
                  },
                  {
                    "priceUpper": 60000,
                    "priceLower": 50000,
                    "count": 76
                  },
                  {
                    "priceUpper": 40000,
                    "priceLower": 30000,
                    "count": 14
                  },
                  {
                    "priceUpper": 50000,
                    "priceLower": 40000,
                    "count": 6
                  },
                  {
                    "priceUpper": 30000,
                    "priceLower": 20000,
                    "count": 1
                  },
                  {
                    "priceUpper": 40000,
                    "priceLower": 30000,
                    "count": 6
                  },
                  {
                    "priceUpper": 50000,
                    "priceLower": 40000,
                    "count": 1
                  },
                  {
                    "priceUpper": 40000,
                    "priceLower": 30000,
                    "count": 2
                  },
                  {
                    "priceUpper": 50000,
                    "priceLower": 40000,
                    "count": 2
                  },
                  {
                    "priceUpper": 60000,
                    "priceLower": 50000,
                    "count": 1
                  }
                ]
              },
              "headers": {
                "Access-Control-Allow-Origin": {
                  "type": "string",
                  "default": "*"
                },
                "Strict-Transport-Security": {
                  "type": "string",
                  "default": "max-age=15552000; includeSubDomains"
                },
                "X-DNS-Prefetch-Control": {
                  "type": "string",
                  "default": "off"
                },
                "X-Frame-Options": {
                  "type": "string",
                  "default": "SAMEORIGIN"
                },
                "X-Download-Options": {
                  "type": "string",
                  "default": "noopen"
                },
                "X-Content-Type-Options": {
                  "type": "string",
                  "default": "nosniff"
                },
                "X-XSS-Protection": {
                  "type": "string",
                  "default": "1; mode=block"
                },
                "Surrogate-Control": {
                  "type": "string",
                  "default": "no-store"
                },
                "Cache-Control": {
                  "type": "string",
                  "default": "no-store, no-cache, must-revalidate, proxy-revalidate"
                },
                "Pragma": {
                  "type": "string",
                  "default": "no-cache"
                },
                "Expires": {
                  "type": "string",
                  "default": "0"
                },
                "Content-Length": {
                  "type": "string",
                  "default": "705"
                },
                "ETag": {
                  "type": "string",
                  "default": "W/\"2c1-I39EyAcvuYCYt5LeZvBy1HQlVms\""
                },
                "Vary": {
                  "type": "string",
                  "default": "Accept-Encoding"
                },
                "Date": {
                  "type": "string",
                  "default": "Wed, 17 Apr 2019 12:40:52 GMT"
                },
                "Connection": {
                  "type": "string",
                  "default": "keep-alive"
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/RegisterVehicleUnauthorised"
              },
              "examples": {
                "application/json": {
                  "name": "AuthorizationRequiredError",
                  "message": "Authorization is required for request on GET /api/histogram/get?year=2013&make=tesla&model=Tesla%20Model%20S&battery=60",
                  "errors": []
                }
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "GetAllScatterPlotSuccess": {
        "title": "GetAllScatterPlotSuccess",
        "example": {
          "km": 33366,
          "originalRange": 98.25
        },
        "type": "object",
        "properties": {
          "km": {
            "example": 33366,
            "type": "integer",
            "format": "int32"
          },
          "originalRange": {
            "example": 98.25,
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "km",
          "originalRange"
        ]
      },
      "GetHistogramSuccess": {
        "title": "GetHistogramSuccess",
        "example": {
          "priceUpper": 50000,
          "priceLower": 40000,
          "count": 26
        },
        "type": "object",
        "properties": {
          "priceUpper": {
            "example": 50000,
            "type": "integer",
            "format": "int32"
          },
          "priceLower": {
            "example": 40000,
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "example": 26,
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "priceUpper",
          "priceLower",
          "count"
        ]
      },
      "GetAllHistogramdataSuccess": {
        "title": "GetAllHistogramdataSuccess",
        "example": {
          "priceUpper": 30000,
          "priceLower": 20000,
          "count": 1
        },
        "type": "object",
        "properties": {
          "priceUpper": {
            "example": 30000,
            "type": "integer",
            "format": "int32"
          },
          "priceLower": {
            "example": 20000,
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "example": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "priceUpper",
          "priceLower",
          "count"
        ]
      },
...
      "RegisterVehicleUnauthorised": {
        "title": "RegisterVehicleUnauthorised",
        "example": {
          "name": "AuthorizationRequiredError",
          "message": "Authorization is required for request on POST /api/vehicle/register",
          "errors": []
        },
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "message",
          "errors"
        ]
      }
    },
...
  }